Cowardice
All the circle-finding algorithm needed was to not think it okay to say "the penumbra must be wayyyy over there. Then we would expect to see nothing, and we did see nothing (because we're looking where the it just pointed). I wrote a routine to solicit user input for finding the center of the penumbra that I don't need, but I'm leaving it commented out because it's cool and I worked hard on it.