Faster!
I was able to make the program substantially more efficient by being more clever about which bins to include in the sums. then I made it slower again by increasing the resolutino. this was awful and I never want to have to use the slice() function again. also, I simplified the plots.