Striking strike!
I wrote a sweeping program to do some statistics scans for proper resolution studies. it's hard to make code work in a lot of different regimes. I had to generalie a bunch of stuff and completely change the way background is estimated when fitting a penumbrum. it sort of works now.