that makes two things in this house
it actually almost works.  it still doesn't quite.  I haff to fix an issue with negative kernel size.  but after that, I think this is goden.  I rerote a lot of the 1D fits to ensure that it is continuous in all parameters.  that meant reriting the included pixel list as linear weits, and improving the interpolacion for the electric feeld stuff.  I also implemented an inicial 2D sweep to help make sure the gorizme finds its penumbrum.  I also improved the plots, fixd a bug where it crashes when you try to run it, rebased a bit more plotting stuff, made it automatically calculate briteness agen.