keep going.  keep going.
I fixd a bug where my contours were all rotated 45° to the left.  also, the fancy resolucion calculacion I implemented for the 1d model last time turnd out to sometimes be horrible, so I replaced it with a constant.  also, I slitely changed the way the theoretical spectrum is calculated during the 1d fit, just to make it that much more like the 2d fit.