Look at it!  Look at it!  Look at it!
I made it look at all of the data instead of the data just around the penumbral image.  it turns out that that problem where it analyzes the blank area of the CR39 so that it can pretend the actual data doesn't exist because signal makes it uncomfortable is a bigger problem than I expected, and my previous efforts to curtail this effect were unsuccessful.  this seems to work fine and isn't too slow, tho.