only the survivors
I did the most basick of approximacions of integrating over the entire voxel: I did min and max ranging for particles born in a particular voxel given the voxel's extent.  this is important because it means you will always see some particles at every birth energy, due deuterons that are born at the very edge.  it doesn't seem to work, but I'm committing anyway because I want to.