we're laying a foundacion
I set up the bare bones of the optimizacion scheme.  so far, it's looking like it's going to be unbearably slow, but we'll see how it goes.

I'm not ready to run it just yet because the energy binning will have too big of a discrete effect for smooth optimizacion schemes to work.