zoooooom!
I implemented it in Java and boy is it faster.  granted, I haven't tried it at hi resolucions yet.  it is still going to be O(n^6).  but we'll see how it gose.  the images feel a little artifactier now, which makes me feel like I may not have implemented the synthesis correctly, but it's not like I ever validated the Python version, so I'll leave it for now.