gradients still suck
but at least the extra complexity is only O(7) instead of O(9)... or something.  I've lost track.  the point is, I don't compute gradients until I need them, like it was originally.  it's better this way.  most of the time the gradient calculacion is pointless.  this way is ~3x faster.