work smarter not harder
I'm done with this multidimensional optimization.  well, mostly.  I've reparameterized it to an actually good parameterization.  now the parameters linearly map to neutron production and density distributions.  I'm still using my Levenberg-Marquardt thing, but it's much more efficient now because the whole problem is _approximately_ linear.  it's still not exactly linear, but I'm realizing that once I parameterize it correctly, it is a lot more linear than for which I gave it credit.  this still isn't incredible in terms of results, but it's better than anything I've had before!