I've done it!
I got a good fit!  I did it!!  I just had to rewrite with a dedicated optimization function and adjust the basis functions to be smoother and now _it works perfectly_!!!
Not only does it work super well, but it's also extremely fast now compared to what I had before.  Time to start ratcheting up the resolucion, baby.
This is great.  I can't wait to apply this to better synthetic data and to real data.