how not to play a private company auction
I set it to precompute all of the complicated stopping power stuff.  it turns out that as long as you assume a uniform temperature and ignore the effects of coupling and of the coulomb logarithm, which I was doing anyway, the Li-Petrasso stopping power is still proportional to density (which makes sense).  so I can have a stopping power curve where all I need is the ρL of each point, which I can compute before I do the full scattering calculation.  and now it's about 10 times faster.