sharing is socialism
I made it bin the things in energy more cleanly, and it doesn't seem to have made a big difference.  there could still be some benefit to having smooth bins, in principle, just to reduce the error in my integrals, but the smoothness for smoothness's sake is no longer necessary.  maybe when I start doing real data for real, I'll implement a fancier line integral along each pixel.

I also adjusted the smoothing so that it doesn't reward big spikes at the outermost radius, but I'm having trubble getting the smoothing to really do anything, so I'm not sure if it works yet.