back to how it was
I changed the entropy weiting parameter back to how it was, and I think what I say it is in the paper, because the sqrt one was not reproducing the decent source reconstructions whose plots I need to modify.

oh, and I also made it output the reconstructed penumbral projection so that I can plot the residual.  that was the main point of all this.