a new paradime
I refactord my code to be much cleaner, more organized, and easier to expand into x-ray image reconstruction.  I haven't added any of that yet, but it's coming.  for now, this is functionally basically the same as what I had before, except that the gelfgat rutine is a bit faster because it's implemented in Java now, and the circle locating algorithm is also faster because it gets the charging parameter from the radius of the 50% contore rather than a forward-fit in 5-space.  unfortunately, the 50% conture thing is not working very well rite now; it seems to be overestimating the charging, leading to bad fits.  I have some ideas, tho.  I think the circle radius, in particular, could be a bit more accurate.  also, I need to just check that my calculation is being consistent about energy bin edges.