can't handle the field
I fixd a bug in the electric field calculation that was causing it to sometimes assume the maximum amount of field rather than fitting it properly (only sometimes, which is why this issue didn't show up in my plots until now).  I also reduced the maximum number of gelfgat iterations and made it give up more politely.