heh, "wiener"
I implemented a Wiener filter reconstruction, mostly just to see what happend.  it actually looks really good!  I need to find a more robust way to set the noise reduction factor, but for now, whatever I've hard-coded is working fine.  this is great because it means I don't haff to rely on that garbage Seguin backprojection, and because now I am really confused as to why the backprojection is so bad.