I made it use more memory
it's faster agen!  there was a relatively simple bug that was making it do the same calculations millions of times.  but now it's reasonable.  but now that it gets to the memorized part of the code, it instantly fails from excessive heap space usage.  I'm hoping that I can fix that by replacing the memory HashMap with an array.