no resampling necessary
I made it so that it keeps the existing resolution instead of resampling when resampling would only change the resolution by a minor factor, so as to reduce degradation of the data due to aliasing.  I was hoping that this was the problem with my Seguin reconstruction, tho this doesn't really seem to have helpd.

I also laid some groundwork for native cpsa support in the future, and simplified the import conventions for math functions, so those are most of the deltas.