implemented an algorithm I will never use
I implemened the gaussian version of gelfgat after realizing that the poisson version is just a better form of richardson-lucy (so I don't need to implement that one), and then I read a little bit about fast convolution and quickly discoverd that I'm a *fool* for using Java for this at all in the first place, because a) the for-loop I was afraid of in Python is actually still convolution so I don't need a for-loop at all if I use Python, and b) scipy has a fftconvolve that is surely just faster than the direct convolution I rote in Java.  so I'm going to bring back the fully-python version now and throw this java file away.