interpolation
I rerote the SphericalHarmonicBasis class to be based on interpolation so that it doesn't haff to calculate a ton of trancendental functions every time it wants to find the value at a spot.  it's like the caching I had before, but not nearlyas memory intensive.  it's still slower than it was before, tho, and I don't know if the interpolation just has too many steps or if it's still sampling more points than necessary for some reason.