I did it!
I made it faster than it was before!  sort of.  it's surprisingly hard to compare, because I've changed so many of the resolution scalings.  but it's actually fast like I would expect now!  well, it still scales pretty badly, but *relatively* it's super fast.  the problem was that I had my resolution set 2x too fine which was making it sample 16x more points.  I also made it look up basis functions more efficiently in SphercialHarmonicBasis.  also, intellij sort of insisted on cleaning up the whitespace.  that didn't affect the speed but it's most of the diff.