snap to grid
I made it fit a hexagonal array of circles to the image to infer the magnification.  it was a little iffy at first, but I realized that in addition to the magnification and array angle, I also really need to fit the array's skewness.  taking the 2-norm of the transformation matrix as the effective magnification seems to work alright, tho it does still kind of look like it's systematicly underestimating the magnification.  may just be an issue with the apertures.