elliptical apertures
I ignored patrick's advice and made it use elliptical apertures.  but listen, it fits *much* better now.  but now it's even more noticeable that the apertures appear larger relative to their spacing.  which is annoying, because that does make it ambiguus what length scale I should use between the source plane and the image plane.  it's only a 0.5% difference so hopefully that aspect doesn't matter.  but it's still weerd...