correct electron temperatures!
the electron temperature fitting was really bad, so I refactored it some.  now it does a 1D optimization where it analyticly sets the prefactor to its optimal value for each given Te.  this takes care of some near-degeneracy with which it was struggling before and lets me quickly converge to good solutions.