line-integrated densities?
I had it output the prefactor so that I can estimate the total emission at any given point (if we could ignore sensitivity).  it’s shockingly irregularly shaped; apparently the total emission you infer is extremely sensitive to temperature (and the tepmerature itself is already pretty sensitive).  so idk if this is useful.  it’s interesting, I gess...