readable code!
I refactored pretty substantially to make it easier to stack images from different parts of the same detector.  I think it makes it much more readable.  it’s exteremly possible that some important part of the code got dropd (like that time I accidentally dropd the part where it uses the "rotation" input).  I’ll be trying to check for that kind of thing this week.  the colors are also a little odd.  but overall I’m quite happy with this.