Absolute corregistration of x-ray images
I made it get the alinement rite between different x-ray images on the same detector, and it turns out there’s a pretty big offset in the high- and low-energy images on each of these detectors, at least in the 1-stalk case.  this is pretty cool, even if it took me way too long to get it rite!