reverse order of method implementations
I’ve decided to reverse my philosophy on the arrangement of methods.   placing the lowest-level methods at the top and the main method at the bottom is a disgusting holdover from the days when the compiler could only read the file once.  now that we have modern programming languages, it’s much better to place them in the order someone would read them: high-level overview at the beginning, and basic routine definitions at the end.  there are no actual code changes in this commit.