you’re far too trusting
I made it trust circles more easily.  I know Patrick told me not to trust them, but now I can actually infer the skew in the array, which is nonzero and has a noticeable effect on the inferred image offset (tho there’s *still* a systematic temperature gradient to the upper left... don’t know what’s up with that).