SRTe
I generalized it to handle SRTe data as well as TIM-based data.  it analyze SRTe *well*, but that’s a task for later.  the biggest change was generalizing the terminology from "TIM" to "LOS".  I also removed support for pickle files and incorporated a more flexible grid shape specification system.