accounting for tilted image plates
it now rotates the data to match the aperture pattern.  this is important for SRTe, because for some reason all of the image plates for it are extremely tilted.  XRIS only seems to have a slite problem, but now I’ve solved that too I gess.

other improvements include changing the order in which the user is shown filtering sections such that it can fit the grid to the big SRTe area first, and hopefully fixing that weerd issue where the 1d reconstruction looks super weerd, and also excluding the one SRTe pinhole that’s misplaced.