clearer output
I added a *slitely* more helpful error message when h is nan, tho of course it can only be so helpful when I don't know what causes it.  I also removed the date from the log messages because, while they could be useful for the log file, they're unnecessary for the stream handler now that I rarely need to run it overnite.