fix 1D reconstruction plots
I fixd several issues with plots because I didn't think the 1d reconstruction was working correctly.  it still doesn't give good ansers, but it does at least create reasonable plots now.  the main issue was that I was plotting a point-spread function with a completely different deuteron energy than that with which I calculated it.  I think the 1d PSF matrix was also weited rong.