add flag for expanding the domain
the last data I analyzed was too big so I had to adjust the code to allow for it.  I don't think it's a good idea to make that the normal procedure so I'm leaving it in this flag that can be set by editing the source.  maybe a command line flag someday if I determine that these parameters actually generally work okay.