the user wasn't talking to you
the user_defined_region function has had some airly janky and annoying functionality for a while.  specificly, if you click at all on the plot, even if it's outside the plotting area so you're not placing a vertex, it will make the default polygon go away, and there's no way to bring it back.  which is misleading because the code will use the default polygon as long as you don't place at least three vertices.  so I fixd that by making it always ignore clicks outside the plot region, and making it so the grey default polygon comes back if you delete your points.