the [] is now optional
if you put the filtering as a list of materials and thicknesses without a [] at the end, it will assume that there's just one detector and that it's at the very end.  I think that's a fair assumption given that most setups these days should just have one detector at the very back.