the [] is now *actually* optional
it was working for a bit but then at the end it would freak out when counting the number of detectors if you didn't include any detectors.  so that's fixd now.  also, I fixd a bug where count_detectors was still using ":" for CR39 and thus would have always returned 0 (yay).