only load each CR39 scan once
when I wrote this code, I was working with pretty big CR39 files exported to text.  I was afraid of having all that track data in memory all the time so I wrote it to load the text file, histogram the tracks, and then immediately forget the tracks.  if I wanted to rebin the histogram, I would load it agen and rehistogram it just to make sure I wasn't losing any image quality.

now I'm using CR39py to read the cpsa files directly, tho, which is way slower, and I'm realizing that having that in memory really isn't so bad as long as only one file is in memory at a time (and even if not I think it's fine).  I gess I have a new laptop now /shrug  anyway, to save time, now it only loads each one once and keeps it around.