catch this edge case
this function is necessary to detect when the number of tracks is zero, but it was failing when precisely that was true.  now there's a check.  hopefully the nans don't cause more problems down the line.  i don't think they are; if the number of tracks is zero it should just quit.