I see your halo
I made it pickier about which contour it uses to measure the P0 and P2 of the source, because a bunch of these are getting that ring artifact just by virtue of the limited iterations, and so the source radius naively looks like the radius of the domain.  but this way it is smart enuff to ignore that contour and look at the main thing in the center.