I finally figured out what that 3 was for
sometimes the colors are based on the detector index, not the cut index, and I approximated the number of detectors as 3.  of course, the error I was getting from removing the 3 would still have been an error had the 3 been kept, if the number of detectors was ever, say, 4.  anyway, now it actually checks the number of detectors, which is just better.