stop interrupting me every five seconds
it now shows all of the plots at the very end.  there was perhaps a time when it made sense to force the user to approve every plot the second it was available, but now that the analysis only takes a minute or so, it's way better to just do the whole thing and show the user the result at the very end, only interrupting when an analysis is done (and therefore might need to be repeated) or when the user's input is required.  this also makes the whole skipping thing just a little bit simpler, tho I suspect it still doesn't quite work.  the issue is that there's so much information in so many different places.  if I really want --skip to be robust I need to have a single analysis object that gets saved to a single HDF5 file like Lotus does.  but I don't think that's a priority now.