the worst possible Markov chain
I created a file where I'm going to put all the MCMC stuff soon, tho for now it just adds some random noise to the image.  I made all the necessary alterations to the code to handle a chain of images rather than just an image, and implemented a 90% credibility interval while I was at it.

broad changes that were made to accomodate this functionality included
- allowing images to have 4D or higher data
- writing a new more efficient resampling function for when we don't have to account for rotation (it's actually copied from last year before I implemented rotation)
- a new Interval class to make it easier to deal with pairs of floats (I think this one should have been a separate commit but oh well)