fix HDF5 saving problem
in the version of the main code I branched this branch out of, there was a bug that prevented the last part of the analysis, where it saves the results, from executing.  that meant that I didn't catch this one error when I implemented the Intervals but it is a problem now that I've merged recent changes from main (because I've since fixd that problem in main).