get correlated!
I implemented spacial correlation of the pixels on an unknown scale.  it's interesting; the resulting images look kind of white-noisy, whereas I expected the noise to be focused at one particular length scale depending on the scaling penalty I applied.  it's interesting.  anyway, I'm finally happy with this prior and think it's time to start looking at data.