likelihood?
I implemented the actual convolution and likelihood part!  but now it's pretty slow.  by which I mean I ran it all evening and nite and it never finished a chain of just 100 samples.  so some further tuning may be in order.  I wonder if I can get anything from simplifying my prior, or if the convolution is just inherently that slow.