make plots more robust
in my quest to achieve a working MCMC I have produced many a bizzare or unphysical source reconstruction.  while I hope to be rid of them before HighTemp, it is important that when such a distribution does arise, the plotting function show how wrong it is rather than failing outrite.  so I've introduced several changes to help prevent such errors.  from clarifying the location of the x-axis on the residual plot to randomizing the samples shown from the chain and allowing a non-error return value from credibility_interval() and more, this makes the plotting functions much more robust to all manner of wrong ansers.