remove infs the moment they form
this kept crashing specificly on the cluster (but not on my laptop).  apparently there were some infs appearing
somewhere in the inverse fourier transform?  I'm attributing this to numerical instability because that's the
only explanation of which I can think that makes sense.  simply clipping any unreasonably extreme values should
hopefully fix that.