tighter bayesian contour plots
I now make the outside of each contour the levelth contour of the 5% quantile and the inside the levelth contour of the 95% quantile.  in the limit of infinite resolution, this is equivalent to my previus strategy of making the outside the 5%th contour of the quantile of level and the inside the 95%th contour of the quantile of level.  but because of the way matplotlib linearly interpolates for contour plots combined with the fact that the quantile of level is very not linear, this will give me much titer contours for finite resolution, in accordance with how the data logicly *should* be interpolated, rather than tending to widen the contour to a whole pixel's width.  it does occasionally now also give me inside-out contours.  but I'm choosing to ignore that for now because I do think that's also just a resolution issue.