fix constrained prob-distribution
it turns out for DensityDist (any distribution?) PyMC does not respect the initval kwarg and instead relies on this "support_point" function, which draws from a uniform [-1, 1] distribution for DensityDist.  meaning that it is not possible to make a DensityDist with a domain constraind to positive numbers.  whatever, it's fine.  I ended up changing my densitydist to be on the log of the source pixel values rather than on the pixel values themselves.  that way it does work for negative values but will naturally never produce negative pixel values.  I don't think it slows anything down since I need the logs of the pixel values anyway for the gamma distribution.  anyway, hopefully this sort of works for fixed noise scale?  still need to figure out how to get the distribution right for varying noise scale.