make pixel value scaling better
okay so it's still not using my initial values, meaning that the ideal source needs to have pixel values of order unity.  I really hope I can find a workaround to this at some point but for now, scaling things with the number of pixels in the guess source should fix it.