fix --skip agen
there were issues in general with duplicate rows in dataframes.  it turns out that in pandas, for a single-index dataframe, Dataframe.loc[str] returns a Series if there is exactly one of that index and a Dataframe if it is duplicated, but for a multi-index dataframe, Dataframe.loc[(str, str)] returns a Series if there is exactly one of *every* index and a Dataframe if there are *any* duplicates in the index.  this came up while I was reriting load_shot_info to be more case-insensitive, which was the actual problem in this case.  anyway, --skip works now.  woo.