Update mcmc.py
I added variable smoothing, and I *think* it works!  the math to normalize this was fun; I enjoyd the puzzle of calculating the prefactor in the minimal way possible.  while I was at it I decoupled the guess number of pixels from the RL gess because I felt bad about using the RL guess for anything.