tamp down edges of source
this change should strongly disincentivize the thing where there's a small region of brightness at the edge of the domain that gets interpreted as the real signal.  I hope, anyway.  since this new edge penalty fully constrains the system in the smoothing framework, I no longer need the sum_penalty (in fact, I must remove it if I want the prefactor to still be easily calculable), which is kind of nice since I never liked it.  hopefully this code is correct; as always, it's hard to test.