force positive values
now that I've finally figured out the perfect prior, I can in fact restrict values to the positive sector.  I'm using the same logarithm trick as before.  see, before this wouldn't have normalized correctly with varying smoothing because the shape of the distribution was changing, and there was no way to know how much of the volume of the distribution was in the all-positive sector.  but now that every single term in the precision matrix is directly proportional to the smoothing, it doesn't matter because the exact same *fraction* of the distribution's volume will always be in that sector, so I can trivially predict how the volume in that sector scales with smoothing.