stop using pandas
I hate software developers.  I really do.  deprecating and removing a method in one year just so that newcomers won't write inefficient code just doesn't make any sense to me.  there are already tons of users who *have* code that *works*!  I just want code to work for longer than two years!  why is this such a deprioritized concern among actual devs?

also I think pandas was not the correct tool for this anyway.  but I didn't know how good numpy structured arrays were.  I blame dan boyer.  he told me that pandas was good and that I should use it.  it's so bad, tho!  it has efficiency and features but the user experience is so awful; all of the syntax is so arcane and unituitive and has to be so precise.  I'm really mad about this, in case you couldn't tell.