open up source domain to full square
cropping this to a circle is making things tricky now that I have MCMC, which doesn't do that, so I think it's time to ditch this feature and just let the source use whatever pixels in the image it wants.  this should move background-related artifacts away from the source and lead to fewer questions in general.