improve 17% contour selection
the technique I had been using to identify "the" 17% conture when there were several has been janky for a while now.  this whole "sussiness" thing... it's not worth it.  and it causes problems in MCMC when the source goes outside the circle region.  so I'm getting rid of it and just using the conture with the largest positive area.  cleaner and more robust.