stop unexpectedly changing error settings
it was really bad of me to set the numpy error settings inside this low-level function and then never set them back.  this would maybe be a cool time for a function decorator?  to ensure that the settings always get set back when it's done.  but for now this is probably fine.