stop ceiling reconstructed sources
while it is true that the actual source must be nonnegative at each pixel, the fact is that the Wiener algorithm doesn't take that constraint into account, and simply setting every negative pixel to zero doesn't serve as any kind of reasonable correction.  on the contrary, it totally screws up the source's sum, which then messes up the background calculation, which just results in nonsense ansers and horrible looking residuals.  no, no, better to deal with such constraints as nonnegativity in the individual deconvolution algorithms, and if the idea of negative pixels is such a problem for you, just don't use Wiener.