handle SR-TE grid better
while I did make sure it wasn't snapping that one misplaced aperture to its theoretical location, I had not, until now, dealt with its effect on the grid finding.  as a result, it was throwing off the magnification calculation, which was probably having some bad effect on the temperature profile calculations.