update calculate_temperature.py
this was outdated; lines of sight are no longer lowercased in the table, the reconstructed yield is now an interval rather than a single number.  also, I made it so that it assumes TPS2 if you don't specify a stalk position, and it is better about filling in gaps when not all your shots have the same number of measured energies.