enhance the MCMC algorithm
now there are options for varying the aperture radius (already implemented but now easily disablable) and adding a term that makes it scale invariant in the limit of low data.  the latter thing is a gaussian distribution that goes over the source and whose radius has a ruffly log-uniform distribution so that it doesn't always gess the largest source it can justify.  this is now disabled because it makes it noticeably slower and apparently doesn't really matter now that I'm no longer using the modified point-spread function.

I also made it use coarser resolution when it's doing MCMC and made it not hang indefinitely when it lacks user input.