fix MCMC
this was really bad.  it sometimes worked, but it didn't need to be as slow as it was.  I seem to remember that I tried using initval before and it just didn't work, but I'm thinking it may be because I was using draw(), which I don't think uses initval.  so I may have just commented it out out of some mistaken belief that it wasn't working when it was?  I also didn't know about jittering so that may also have confused me.  in any case, actually passing a reasonable initval makes it work *so* much better.  I can reduce the number of samples by a factor of, like 3.  I also raised target_accept because 80% is a bizzarely low acceptance rate in my onion.