manual alignment for calculating Te
previusly it would always assume that the images that come out of reconstruct_2d are correctly aligned with each other.  this is not always the case.  that alinement is based on fitting the known aperture arrangement to the data, but if the actual aperture positions differ from that at all, you can get systematic shifts that offset the sources relative to each other.  this is what we see on the particular shot day I'm analyzing.  the BELIEVE_IN_APERTURE_TILTING option helps, but doesn't generally go far enuff to eliminate the offset.  I beleve this is most likely because the apertures were not just tilted but actually bent on that shot day.  hopefully this doesn't happen anymore in the future, but in case it does, there's an option in the calculate_temperature.py source now.