reduce free parameters for MCMC
I realized that my MCMC was still too slow at resolutions finer than 5μm (which is really not much), and that it's largely because I have too many input parameters.  the vast majority of the pixels in the source are zero, and are known to be zero ahead of time, so it didn't make any sense to include them as free parameters.  it took me a while to find out how to set subtensors in pytensor, but once I did it was relatively easy to make it only vary the pixels that have significant values in the gelfgat reconstruction.  I'm trying to be conservative here, so I still probably vary more than is necessary, but this makes the problem tractable and doesn't seem to affect the result, so I'm pretty happy with that.  also, more generally, requiring that the source domain always be 500μm across was kind of insane, so fixing that also sped things up a lot.