remove restrictions on source size
Tucker wants to use this code to analyze some really big sources, so I turned the max source size up to 1 mm.  several adjustments were needed to accomodate this in a reasonable manner.  the number of pixels is now fixed instead of the pixel size by default – much more logical imho.  the algorithm used to determine the domain size for the 1D reconstruction is totally different because I don't really understand why it was the way it was.  now it's just the distance to the point between apertures, or 5cm.  I think that should work well enuff for both single apertures and arrays.  I also removed FORCE_LARGE_SOURCE_DOMAIN because it should no longer be necessary.